delete class DashboardBreadcrumbsTest delete class DashboardLineTest and moved tests to DashboardTest
